refactor: scaffold python_api package
